fix typo and fix bug

+++ installer/install-fonts.sh
@@ -8,13 +8,17 @@ function install_fonts() {
   command echo -e "Running... $(basename $0)"
   command echo ""
   local distro=$(whichdistro)
+  local current_dir="$(cd "$(dirname "$0")" && pwd)"
   if [[ $distro == "arch" ]]; then
     yay -S --noconfirm --needed noto-fonts-cjk ttf-jetbrains-mono-nerd \
       noto-fonts-emoji ipa-fonts
-    cd /tmp/
-    git clone https://github.com/coz-m/MPLUS_FONTS.git
-    mv MPLUS_FONTS/fonts/ttf/* /usr/share/fonts/TTF/
-    rm -rf MPLUS_FONTS
+    if ! ls /usr/share/fonts/TTF/ | grep -qE '^(Mplus|MPLUS)'; then
+      cd /var/tmp
+      git clone https://github.com/coz-m/MPLUS_FONTS.git
+      sudo mv MPLUS_FONTS/fonts/ttf/* /usr/share/fonts/TTF/
+      rm -rf MPLUS_FONTS
+      cd $current_dir
+    fi
   elif [[ $distro == "ubuntu" ]]; then
     :
   fi
